+++ Composer/composer.py
@@ -3,7 +3,7 @@ from Data import ExperimentDescriptor, ExperimentType, NsInfo, Metal
 from .platform_configuration import PlatformConfiguration, TaskDefinition
 from importlib import import_module
 from Helper import Log
-from typing import List
+from typing import List, Dict, Optional, Tuple
 from Executor.Tasks.Run import Message
 from Interfaces import Management
 
@@ -34,41 +34,54 @@ class Composer:
 
         actions: List[ActionInformation] = []
         panels: List[DashboardPanel] = []
+        errored = False
 
-        if len(descriptor.NetworkServices) != 0:
-            sliceManager = Management.SliceManager()
-            for ns in descriptor.NetworkServices:
-                nsId, location = ns
-                try:
-                    nsInfo = NsInfo(nsId, location)
-                    requirements = sliceManager.GetNsdRequirements(nsId)
-                    if requirements is None:
-                        raise RuntimeError("Could not retrieve NSD information")
-                    nsInfo.Requirements = requirements
-                    configuration.NetworkServices.append(nsInfo)
-                except Exception as e:
-                    actions.append(
-                        _messageAction("ERROR",
-                                       f"Exception while obtaining information about network service {nsId}: {e}"))
+        if descriptor.Slice is not None:
+            if len(descriptor.NetworkServices) != 0:
+                sliceManager = Management.SliceManager()
+                for ns in descriptor.NetworkServices:
+                    nsId, location = ns
+                    try:
+                        nsInfo = NsInfo(nsId, location)
+                        requirements = sliceManager.GetNsdRequirements(nsId)
+                        if requirements is None:
+                            raise RuntimeError("Could not retrieve NSD information")
+                        nsInfo.Requirements = requirements
+                        configuration.NetworkServices.append(nsInfo)
+                    except Exception as e:
+                        errored = True
+                        actions.append(
+                            _messageAction("ERROR",
+                                           f"Exception while obtaining information about network service {nsId}: {e}"))
 
-        if descriptor.Type == ExperimentType.MONROE:
-            actions.extend(cls.facility.GetMonroeActions())
-        else:
-            if descriptor.Automated:
-                for ue in descriptor.UEs:
-                    actions.extend(cls.facility.GetUEActions(ue))
-                for testcase in descriptor.TestCases:
-                    testcaseActions = cls.facility.GetTestCaseActions(testcase)
-                    if len(testcaseActions) != 0:
-                        actions.extend(testcaseActions)
-                    else:
-                        actions.append(_messageAction("ERROR", f'TestCase "{testcase}" did not generate any actions'))
-                    panels.extend(cls.facility.GetTestCaseDashboards(testcase))
+            if not errored:
+                nest, error = cls.composeNest(descriptor.Slice, descriptor.Scenario, configuration.NetworkServices)
+                if error is None:
+                    configuration.Nest = nest
+                else:
+                    errored = True
+                    actions.append(_messageAction("ERROR", f'Error while generating NEST data for experiment: {error}'))
+
+        if not errored:
+            if descriptor.Type == ExperimentType.MONROE:
+                actions.extend(cls.facility.GetMonroeActions())
             else:
-                delay = ActionInformation()
-                delay.TaskName = "Run.Delay"
-                delay.Config = {'Time': descriptor.Duration*60}
-                actions.append(delay)
+                if descriptor.Automated:
+                    for ue in descriptor.UEs:
+                        actions.extend(cls.facility.GetUEActions(ue))
+                    for testcase in descriptor.TestCases:
+                        testcaseActions = cls.facility.GetTestCaseActions(testcase)
+                        if len(testcaseActions) != 0:
+                            actions.extend(testcaseActions)
+                        else:
+                            actions.append(_messageAction("WARNING",  # Notify, but do not cancel execution
+                                                          f'TestCase "{testcase}" did not generate any actions'))
+                        panels.extend(cls.facility.GetTestCaseDashboards(testcase))
+                else:
+                    delay = ActionInformation()
+                    delay.TaskName = "Run.Delay"
+                    delay.Config = {'Time': descriptor.Duration*60}
+                    actions.append(delay)
 
         actions.sort(key=lambda action: action.Order)  # Sort by Order
         requirements = set()
@@ -99,3 +112,35 @@ class Composer:
         except (ModuleNotFoundError, AttributeError, ValueError):
             Log.E(f'Task "{taskName}" not found')
             return None
+
+    @classmethod
+    def composeNest(cls, baseSlice: str, scenario: str, nss: List[NsInfo]) -> Tuple[Dict, Optional[str]]:
+        """Returns the NEST as Dict (possibly empty) and a string indicating the error (None in case of success)"""
+        try:
+            if baseSlice is None:
+                raise RuntimeError("Cannot create NEST without a base slice value")
+
+            nsList = []
+            for ns in nss:
+                nsList.append({
+                    "nsd-id": ns.Id,
+                    "placement": ns.Location,
+                })
+
+            sliceDescriptor = {"base_slice_des_id": baseSlice}
+
+            if scenario is not None:
+                scenarioData = Facility.Scenarios().get(scenario, None)
+                if scenarioData is None:
+                    raise RuntimeError(f"Unrecognized scenario '{scenario}'")
+                sliceDescriptor.update(scenarioData)
+            else:
+                raise RuntimeError("Cannot create NEST without an scenario value")
+
+            nest = {"base_slice_descriptor": sliceDescriptor}
+            if len(nsList) != 0:
+                nest["service_descriptor"] = {"ns_list": nsList}
+
+            return nest, None
+        except Exception as e:
+            return {}, str(e)