+++ Executor/Tasks/PostRun/decommission.py
@@ -13,14 +13,21 @@ class Decommission(Task):
         self.Log(Level.INFO, 'Decommision started')
         networkServices: List[NsInfo] = self.params['NetworkServices']
         if len(networkServices) != 0:
-            self.Log(Level.INFO, f"Experiment has {len(networkServices)} network services")
-            for ns in networkServices:
-                self.Log(Level.INFO, f"Requesting decommision of NS {ns.Id} with slice ID: {ns.SliceId}")
+            sliceId = networkServices[0].SliceId  # All NS should share the same slice ID
+            if sliceId is not None:
+                self.Log(Level.INFO, f"Experiment has {len(networkServices)} network services with slice ID: {sliceId}."
+                                     f"Requesting decommision")
                 try:
-                    # TODO
-                    self.Log(Level.INFO, f'Slice decommisioned')
+                    message, success = Management.SliceManager().DeleteSlice(sliceId)
+                    if success:
+                        self.Log(Level.INFO, f'Slice decommisioned')
+                    else:
+                        self.Log(Level.ERROR, message)
                 except Exception as e:
                     raise Exception(f'Exception while decommisioning slice: {e}') from e
+            else:
+                self.Log(Level.WARNING, 'Experiment contains network services, but no slice ID. '
+                                        'Please review the Pre-Run log for instantiation issues.')
         else:
-            self.Log(Level.INFO, 'Decommision not required, no NSD IDs defined.')
+            self.Log(Level.INFO, 'Decommision not required, no Network Services defined.')
         self.Log(Level.INFO, 'Decommision completed')