+++ Executor/Tasks/PreRun/instantiate.py
@@ -1,31 +1,30 @@
 from Task import Task
 from Helper import Level, Config
 from Data import NsInfo
-from typing import List
+from typing import List, Dict
+from Interfaces import Management
+from json import dumps
 
 
 class Instantiate(Task):
-    def __init__(self, logMethod, tempFolder, parent, networkServices):
-        super().__init__("Instantiate", parent, {'NetworkServices': networkServices}, logMethod, None)
+    def __init__(self, logMethod, tempFolder, parent, networkServices, nest):
+        super().__init__("Instantiate", parent, {'NetworkServices': networkServices, 'NEST': nest}, logMethod, None)
         self.tempFolder = tempFolder
 
     def Run(self):
-        networkServices: List[NsInfo] = self.params['NetworkServices']  # TODO: this is List[NsInfo] now
-        sliceIds = []
+        networkServices: List[NsInfo] = self.params['NetworkServices']
+        nest: Dict = self.params['NEST']
 
         if len(networkServices) != 0:
-            self.Log(Level.INFO, f"Experiment contains {len(networkServices)} NSD IDs")
-            for ns in networkServices:
-                self.Log(Level.INFO, f"Requesting instantiation of NSD: {ns.Id}")
-                try:
-                    # sliceId = Management.SliceManager().CreateSlice(nsdContent)
-                    sliceId = "placeholder"  # TODO
-                    self.Log(Level.INFO, f'Network service instantiated with ID: {sliceId}')
-                    ns.SliceId = sliceId
-                except Exception as e:
-                    raise Exception(f'Exception while creating slice: {e}') from e
+            self.Log(Level.INFO, f"Experiment contains {len(networkServices)} NSD IDs. Requesting instantiation.")
+            self.Log(Level.DEBUG, f"NEST: '{nest}'")
+            maybeSliceId, success = Management.SliceManager().CreateSlice(dumps(nest))
+            if success:
+                for ns in networkServices:
+                    ns.SliceId = maybeSliceId
+            else:
+                self.Log(Level.ERROR, maybeSliceId)
         else:
             self.Log(Level.INFO, 'Instantiation not required, no NSD IDs defined.')
 
         self.Log(Level.INFO, 'Instantiation completed')
-        self.params["SliceIds"] = sliceIds